issues/5 refactor(rename): UnicacheCore -> Unicache, UnicachePlugin. Unicache -> FileCache
