+++ Assets/Plugins/Unicache/Scripts/Plugin.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 605c564ae6bf64001a6296a85590c9a5
+folderAsset: yes
+timeCreated: 1483895374
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
